fixed a bug in the xmodem code
The destination pointer was never being incremented as data was being
loaded, instead relying on the Y register as an index. This only worked
if the data was 1 packet (128 bytes). Once we exceeded that limit,
subsequent packets would overwrite the same 128 bytes in memory.

This change increments the 16-bit zero-page pointer directly instead of
using the Y register.